# Random-value generation clean-up, and move aux data in client context to separate ClientData class