--- a/src/common/random_engine.cpp
+++ b/src/common/random_engine.cpp
@@ -1,0 +1,43 @@
+#include "duckdb/common/random_engine.hpp"
+#include "pcg_random.hpp"
+#include <random>
+
+namespace duckdb {
+
+struct RandomState {
+	RandomState() {
+	}
+
+	pcg32 pcg;
+};
+
+RandomEngine::RandomEngine(int64_t seed) : random_state(make_unique<RandomState>()) {
+	if (seed < 0) {
+		random_state->pcg.seed(pcg_extras::seed_seq_from<std::random_device>());
+	} else {
+		random_state->pcg.seed(seed);
+	}
+}
+
+RandomEngine::~RandomEngine() {
+
+}
+
+double RandomEngine::NextRandom(double min, double max) {
+	D_ASSERT(max >= min);
+	return min + (NextRandom() * (max - min));
+}
+
+double RandomEngine::NextRandom() {
+	return random_state->pcg() / double(std::numeric_limits<uint32_t>::max());
+}
+uint32_t RandomEngine::NextRandomInteger() {
+	return random_state->pcg();
+}
+
+void RandomEngine::SetSeed(uint32_t seed) {
+	random_state->pcg.seed(seed);
+}
+
+
+}