--- a/src/execution/operator/helper/physical_prepare.cpp
+++ b/src/execution/operator/helper/physical_prepare.cpp
@@ -1,5 +1,5 @@
 #include "duckdb/execution/operator/helper/physical_prepare.hpp"
-#include "duckdb/main/client_context.hpp"
+#include "duckdb/main/client_data.hpp"
 
 namespace duckdb {
 
@@ -8,7 +8,7 @@ void PhysicalPrepare::GetData(ExecutionContext &context, DataChunk &chunk, Globa
 	auto &client = context.client;
 
 	// store the prepared statement in the context
-	client.prepared_statements[name] = prepared;
+	ClientData::Get(client).prepared_statements[name] = prepared;
 }
 
 } // namespace duckdb