--- a/src/function/scalar/math/random.cpp
+++ b/src/function/scalar/math/random.cpp
@@ -3,19 +3,18 @@
 #include "duckdb/execution/expression_executor.hpp"
 #include "duckdb/main/client_context.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
-#include <random>
+#include "duckdb/common/random_engine.hpp"
 
 namespace duckdb {
 
 struct RandomBindData : public FunctionData {
 	ClientContext &context;
-	std::uniform_real_distribution<double> dist;
 
-	RandomBindData(ClientContext &context, std::uniform_real_distribution<double> dist) : context(context), dist(dist) {
+	RandomBindData(ClientContext &context) : context(context) {
 	}
 
 	unique_ptr<FunctionData> Copy() override {
-		return make_unique<RandomBindData>(context, dist);
+		return make_unique<RandomBindData>(context);
 	}
 };
 
@@ -26,15 +25,15 @@ static void RandomFunction(DataChunk &args, ExpressionState &state, Vector &resu
 
 	result.SetVectorType(VectorType::FLAT_VECTOR);
 	auto result_data = FlatVector::GetData<double>(result);
+	auto &random_engine = RandomEngine::Get(info.context);
 	for (idx_t i = 0; i < args.size(); i++) {
-		result_data[i] = info.dist(info.context.random_engine);
+		result_data[i] = random_engine.NextRandom(0, 1);
 	}
 }
 
 unique_ptr<FunctionData> RandomBind(ClientContext &context, ScalarFunction &bound_function,
                                     vector<unique_ptr<Expression>> &arguments) {
-	std::uniform_real_distribution<double> dist(0, 1);
-	return make_unique<RandomBindData>(context, dist);
+	return make_unique<RandomBindData>(context);
 }
 
 void RandomFun::RegisterFunction(BuiltinFunctions &set) {