--- a/src/function/scalar/math/random.cpp
+++ b/src/function/scalar/math/random.cpp
@@ -29,7 +29,6 @@ struct RandomLocalState : public FunctionData {
 
 static void RandomFunction(DataChunk &args, ExpressionState &state, Vector &result) {
 	D_ASSERT(args.ColumnCount() == 0);
-	auto &func_expr = (BoundFunctionExpression &)state.expr;
 	auto &lstate = (RandomLocalState &)*ExecuteFunctionState::GetFunctionState(state);
 
 	result.SetVectorType(VectorType::FLAT_VECTOR);