--- a/src/function/scalar/system/system_functions.cpp
+++ b/src/function/scalar/system/system_functions.cpp
@@ -1,6 +1,7 @@
 #include "duckdb/catalog/catalog_search_path.hpp"
 #include "duckdb/function/scalar/generic_functions.hpp"
 #include "duckdb/main/database.hpp"
+#include "duckdb/main/client_context.hpp"
 #include "duckdb/main/client_data.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/catalog/catalog_entry/aggregate_function_catalog_entry.hpp"