--- a/src/function/table/pragma_last_profiling_output.cpp
+++ b/src/function/table/pragma_last_profiling_output.cpp
@@ -3,6 +3,7 @@
 #include "duckdb/common/limits.hpp"
 #include "duckdb/function/table/system_functions.hpp"
 #include "duckdb/main/client_context.hpp"
+#include "duckdb/main/client_data.hpp"
 #include "duckdb/planner/constraints/bound_not_null_constraint.hpp"
 #include "duckdb/main/query_profiler.hpp"
 
@@ -70,8 +71,8 @@ static void PragmaLastProfilingOutputFunction(ClientContext &context, const Func
 		DataChunk chunk;
 		chunk.Initialize(data.types);
 		int operator_counter = 1;
-		if (!context.query_profiler_history->GetPrevProfilers().empty()) {
-			for (auto op : context.query_profiler_history->GetPrevProfilers().back().second->GetTreeMap()) {
+		if (!ClientData::Get(context).query_profiler_history->GetPrevProfilers().empty()) {
+			for (auto op : ClientData::Get(context).query_profiler_history->GetPrevProfilers().back().second->GetTreeMap()) {
 				SetValue(chunk, chunk.size(), operator_counter++, op.second->name, op.second->info.time,
 				         op.second->info.elements, " ");
 				chunk.SetCardinality(chunk.size() + 1);