--- a/src/function/table/system/duckdb_constraints.cpp
+++ b/src/function/table/system/duckdb_constraints.cpp
@@ -74,7 +74,7 @@ unique_ptr<FunctionOperatorData> DuckDBConstraintsInit(ClientContext &context, c
 	};
 
 	// check the temp schema as well
-	context.temporary_objects->Scan(context, CatalogType::TABLE_ENTRY,
+	ClientData::Get(context).temporary_objects->Scan(context, CatalogType::TABLE_ENTRY,
 	                                [&](CatalogEntry *entry) { result->entries.push_back(entry); });
 	return move(result);
 }