--- a/src/function/table/system/duckdb_sequences.cpp
+++ b/src/function/table/system/duckdb_sequences.cpp
@@ -70,7 +70,7 @@ unique_ptr<FunctionOperatorData> DuckDBSequencesInit(ClientContext &context, con
 	};
 
 	// check the temp schema as well
-	context.temporary_objects->Scan(context, CatalogType::SEQUENCE_ENTRY,
+	ClientData::Get(context).temporary_objects->Scan(context, CatalogType::SEQUENCE_ENTRY,
 	                                [&](CatalogEntry *entry) { result->entries.push_back(entry); });
 	return move(result);
 }