… benchmark/micro/functions/random.benchmark
… benchmark/micro/functions/random_uuid.benchmark
… src/common/CMakeLists.txt
… src/function/scalar/CMakeLists.txt
… src/function/scalar/uuid/CMakeLists.txt
… src/function/table/system/duckdb_tables.cpp
… src/function/table/system/duckdb_views.cpp
… src/include/duckdb/common/cycle_counter.hpp
… src/include/duckdb/common/random_engine.hpp
… src/include/duckdb/execution/expression_executor_state.hpp
… src/include/duckdb/main/client_context.hpp
… src/include/duckdb/main/client_data.hpp
… src/main/CMakeLists.txt
… src/main/client_context.cpp
… src/main/client_data.cpp
… src/main/query_profiler.cpp
… src/main/settings/settings.cpp
… src/parallel/executor.cpp
… src/planner/binder/statement/bind_create.cpp
… src/planner/planner.cpp
… test/sql/window/test_window_wide_frame.test_coverage