This PR cleans up the way in which random number generation is done by moving everything to the general `RandomEngine` class, and by converting the `RandomEngine` to use the `PCG-Random` generator (that we were already using before for samples).

We also clean-up the ClientContext class by moving the included random engine and several other auxiliary variables to a dedicated `ClientData` class. 

Moving the generic random number generator to `PCG-Random` significantly speeds up random number generation. In addition, we rework the `random` and `gen_random_uuid` functions so that the  same random number generator is not shared across threads, which reduces contention and increases generation speed:

|    Benchmark     | Old (s) | New (s) |
|------------------|--------:|--------:|
| Random (1T)      | 0.88    | 0.400   |
| Random (8T)      | 0.24    | 0.058   |
| Random UUID (1T) | 2.23    | 1.180   |
| Random UUID (8T) | 0.67    | 0.170   |

Another advantage of switching to PCG-Random is that random number generation with a specific seed is consistent between different compilations of DuckDB, even across different platforms/standard library versions.